mcp/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
mcp==1.1.2
pydantic==2.10.3
python-dotenv==1.0.1
pyyaml==6.0.2
httpx==0.27.2
sse-starlette==2.1.3
